+++ pmd-core/src/main/java/net/sourceforge/pmd/cache/FileAnalysisCache.java
@@ -74,6 +74,14 @@ public class FileAnalysisCache extends AbstractAnalysisCache {
 
     @Override
     public void persist() {
+        // Create directories missing along the way
+        if (!cacheFile.exists()) {
+            final File parentFile = cacheFile.getAbsoluteFile().getParentFile();
+            if (parentFile != null && !parentFile.exists()) {
+                parentFile.mkdirs();
+            }
+        }
+
         try (
             DataOutputStream outputStream = new DataOutputStream(
                 new BufferedOutputStream(new FileOutputStream(cacheFile)));